Remove unnecessary partial file
